% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarizeByGMM}
\alias{binarizeByGMM}
\title{Binarize numeric vector using GMM}
\usage{
binarizeByGMM(score_vector)
}
\arguments{
\item{score_vector}{A numeric vector}
}
\value{
Integer vector (0/1)
}
\description{
Binarize numeric vector using GMM
}
