% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_by_main_valley}
\alias{classify_by_main_valley}
\title{Classify by main valley threshold}
\usage{
classify_by_main_valley(score_vector, plot = FALSE)
}
\arguments{
\item{score_vector}{Numeric vector}

\item{plot}{Whether to show density plot}
}
\value{
Integer vector (0/1)
}
\description{
Classify by main valley threshold
}
