% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{computePairwiseCelltypeOverlap}
\alias{computePairwiseCelltypeOverlap}
\title{Compute pairwise overlap between cell types}
\usage{
computePairwiseCelltypeOverlap(
  object,
  metric = c("dice", "jaccard", "mcc"),
  ncore = NULL
)
}
\arguments{
\item{object}{A SparrotObj object}

\item{metric}{One of "dice", "jaccard", or "mcc"}

\item{ncore}{Number of cores to use (set to NULL or 1 for sequential execution)}
}
\value{
A data.frame with celltype pairs, overlap values, and p-values
}
\description{
Compute pairwise overlap between cell types
}
