% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{concave_dat}
\alias{concave_dat}
\title{Compute concave hull of coordinate set}
\usage{
concave_dat(coords, concavity = 2)
}
\arguments{
\item{coords}{Matrix of 2D coordinates}

\item{concavity}{Concavity parameter (lower = smoother)}
}
\value{
A data.frame of polygon coordinates
}
\description{
Compute concave hull of coordinate set
}
