% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{convertSeuratToSparrot}
\alias{convertSeuratToSparrot}
\title{Convert a Seurat object to a SparrotObj}
\usage{
convertSeuratToSparrot(
  seurat_obj,
  cell_prob,
  expr_slot = "data",
  assay = NULL,
  params = list()
)
}
\arguments{
\item{seurat_obj}{A Seurat object with spatial coordinates}

\item{cell_prob}{A matrix of cell type probabilities}

\item{expr_slot}{Expression slot to extract (e.g., "data")}

\item{assay}{Assay name (defaults to DefaultAssay())}

\item{params}{A list of additional parameters}
}
\value{
A SparrotObj
}
\description{
Convert a Seurat object to a SparrotObj
}
