% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{createSparrotObj}
\alias{createSparrotObj}
\title{Create a SparrotObj object}
\usage{
createSparrotObj(coords, expr, cell_prob, meta.data = NULL, params = list())
}
\arguments{
\item{coords}{A matrix of spatial coordinates (row, col).}

\item{expr}{A gene expression matrix (dgCMatrix).}

\item{cell_prob}{A numeric matrix of cell type probabilities.}

\item{meta.data}{Optional metadata as a DataFrame.}

\item{params}{A list of additional parameters.}
}
\value{
An object of class SparrotObj
}
\description{
Create a new SparrotObj containing coordinates, expression, cell type probabilities, and metadata.
}
