% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{evaluate_bidirectional_nn_distance}
\alias{evaluate_bidirectional_nn_distance}
\title{Compute bidirectional nearest-neighbor distances between two binary groups}
\usage{
evaluate_bidirectional_nn_distance(bin1, bin2, coords)
}
\arguments{
\item{bin1}{Logical vector}

\item{bin2}{Logical vector}

\item{coords}{Matrix of spatial coordinates}
}
\value{
A list with distance vectors and summary statistics
}
\description{
Compute bidirectional nearest-neighbor distances between two binary groups
}
