% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_bin_chebyshev}
\alias{expand_bin_chebyshev}
\title{Expand binarized region by Chebyshev distance}
\usage{
expand_bin_chebyshev(bin_vec, coords, max_bin_dist = 1)
}
\arguments{
\item{bin_vec}{A logical vector}

\item{coords}{A matrix of spatial coordinates}

\item{max_bin_dist}{Maximum Chebyshev distance to expand}
}
\value{
A logical vector of expanded region
}
\description{
Expand binarized region by Chebyshev distance
}
