% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_main_valley_threshold}
\alias{find_main_valley_threshold}
\title{Find threshold between main valley in density}
\usage{
find_main_valley_threshold(score_vector)
}
\arguments{
\item{score_vector}{Numeric vector}
}
\value{
Numeric threshold
}
\description{
Find threshold between main valley in density
}
