% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellType}
\alias{plotCellType}
\title{Plot binarized presence of a cell type}
\usage{
plotCellType(object, celltype, outline = FALSE, pt.size = 1, color = "red")
}
\arguments{
\item{object}{SparrotObj}

\item{celltype}{Cell type to plot}

\item{outline}{Whether to add outline}

\item{pt.size}{Point size}

\item{color}{Color for present cells}
}
\value{
A ggplot object
}
\description{
Plot binarized presence of a cell type
}
