% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellTypeProb}
\alias{plotCellTypeProb}
\title{Plot probability of a single cell type}
\usage{
plotCellTypeProb(object, celltype, outline = FALSE, pt.size = 1)
}
\arguments{
\item{object}{SparrotObj}

\item{celltype}{Cell type to plot}

\item{outline}{Whether to add concave outline}

\item{pt.size}{Point size}
}
\value{
A ggplot object
}
\description{
Plot probability of a single cell type
}
