% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMultiCellTypeProb}
\alias{plotMultiCellTypeProb}
\title{Plot multiple cell type probabilities with normalized alpha+color blending}
\usage{
plotMultiCellTypeProb(
  object,
  celltype = NULL,
  pt.size = 1,
  outline = TRUE,
  concavity = 2,
  color = NULL,
  coord.fixed = TRUE,
  legend = TRUE,
  normalized = TRUE
)
}
\arguments{
\item{object}{SparrotObj}

\item{celltype}{Cell types to plot}

\item{pt.size}{Point size}

\item{outline}{Whether to add outline}

\item{color}{Optional named color map}

\item{coord.fixed}{Fix aspect ratio}

\item{legend}{Whether to add legend}

\item{normalized}{Normalize probability values to 0-1}
}
\value{
A ggplot object or cowplot
}
\description{
Plot multiple cell type probabilities with normalized alpha+color blending
}
