% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spFeatureDensityPlot}
\alias{spFeatureDensityPlot}
\title{Plot spatial feature density using Nebulosa}
\usage{
spFeatureDensityPlot(
  object,
  features,
  outline = TRUE,
  raster = TRUE,
  pt.size = 1,
  method = c("ks", "wkde"),
  joint = FALSE,
  ncol = NULL
)
}
\arguments{
\item{object}{A SparrotObj}

\item{features}{Vector of gene names}

\item{outline}{Add concave outline}

\item{raster}{Rasterize points}

\item{pt.size}{Point size}

\item{method}{Kernel method ("ks", "wkde")}

\item{joint}{Joint density}

\item{ncol}{Number of columns for multiple plots}
}
\value{
A ggplot object
}
\description{
Plot spatial feature density using Nebulosa
}
